Code cleanup, move ViewHolder classes into same file as their respective activities, since they're not used elsewhere
