+++ android/src/main/java/org/ligi/blexplorer/scan/DeviceListActivity.kt
@@ -1,5 +1,6 @@
 package org.ligi.blexplorer.scan
 
+import android.app.Activity
 import android.app.AlertDialog
 import android.bluetooth.BluetoothAdapter
 import android.bluetooth.BluetoothDevice
@@ -9,18 +10,24 @@ import android.content.DialogInterface
 import android.content.Intent
 import android.os.Bundle
 import android.os.Handler
-import androidx.appcompat.app.AppCompatActivity
-import androidx.recyclerview.widget.LinearLayoutManager
-import androidx.recyclerview.widget.RecyclerView
+import android.text.TextUtils
 import android.view.LayoutInflater
 import android.view.Menu
 import android.view.MenuItem
 import android.view.ViewGroup
+import androidx.appcompat.app.AppCompatActivity
+import androidx.recyclerview.widget.RecyclerView
+import org.ligi.blexplorer.App
 import org.ligi.blexplorer.HelpActivity
 import org.ligi.blexplorer.R
 import org.ligi.blexplorer.databinding.ActivityWithRecyclerBinding
 import org.ligi.blexplorer.databinding.ItemDeviceBinding
+import org.ligi.blexplorer.services.DeviceServiceExploreActivity
+import org.ligi.blexplorer.util.DevicePropertiesDescriber
+import org.ligi.blexplorer.util.ManufacturerRecordParserFactory
+import org.ligi.blexplorer.util.from_lollipop.ScanRecord
 import org.ligi.tracedroid.sending.TraceDroidEmailSender
+import java.math.BigInteger
 import java.util.*
 
 class DeviceListActivity : AppCompatActivity() {
@@ -125,6 +132,61 @@ class DeviceListActivity : AppCompatActivity() {
     companion object {
 
         private val REQUEST_ENABLE_BT = 2300
+    }
+
+}
+
+private class DeviceViewHolder(private val binding: ItemDeviceBinding) : RecyclerView.ViewHolder(binding.root) {
+
+    lateinit var device: BluetoothDevice
+
+    fun applyDevice(newDevice: BluetoothDevice, extras: DeviceListActivity.DeviceExtras) {
+        device = newDevice
+        binding.name.text = if (TextUtils.isEmpty(device.name)) "no name" else device.name
+        binding.rssi.text = "${extras.rssi}db"
+        binding.lastSeen.text = "" + (System.currentTimeMillis() - extras.last_seen) / 1000 + "s"
+        binding.address.text = device.address
+
+        val scanRecord = ScanRecord.parseFromBytes(extras.scanRecord)
+        var scanRecordStr = ""
+        if (scanRecord.serviceUuids != null) {
+            for (parcelUuid in scanRecord.serviceUuids) {
+                scanRecordStr += parcelUuid.toString() + "\n"
+            }
+        }
+
+        val manufacturerSpecificData = scanRecord.manufacturerSpecificData
+
+        (0 until manufacturerSpecificData.size())
+                .map { manufacturerSpecificData.keyAt(it) }
+                .forEach {key ->
+                    val p = ManufacturerRecordParserFactory.parse(key, manufacturerSpecificData.get(key), device)
+                    if (p == null) {
+                        scanRecordStr += "$key=" + BigInteger(1, manufacturerSpecificData.get(key)).toString(16) + "\n"
+                    } else {
+                        scanRecordStr += p.keyDescriptor + " = {\n" + p.toString() + "}\n"
+                        if (!TextUtils.isEmpty(p.getName(device))) {
+                            binding.name.text = p.getName(device)
+                        }
+                    }
+                }
+
+        for (parcelUuid in scanRecord.serviceData.keys) {
+            scanRecordStr += "$parcelUuid=" + BigInteger(1, scanRecord.serviceData[parcelUuid]).toString(16) + "\n"
+        }
+
+        binding.scanRecord.text = scanRecordStr
+
+        binding.type.text = DevicePropertiesDescriber.describeType(device)
+        binding.bondstate.text = DevicePropertiesDescriber.describeBondState(device)
+    }
+
+    fun installOnClickListener(activity: Activity) {
+        itemView.setOnClickListener {
+            val intent = Intent(activity, DeviceServiceExploreActivity::class.java)
+            App.device = device
+            activity.startActivity(intent)
+        }
     }
 
 }
