--- android/src/main/java/org/ligi/blexplorer/services/ServiceViewHolder.kt
@@ -1,22 +1,0 @@
-package org.ligi.blexplorer.services
-
-import android.bluetooth.BluetoothGattService
-import androidx.recyclerview.widget.RecyclerView
-import org.ligi.blexplorer.App
-import org.ligi.blexplorer.characteristics.CharacteristicActivity
-import org.ligi.blexplorer.databinding.ItemServiceBinding
-import org.ligi.blexplorer.util.DevicePropertiesDescriber
-import org.ligi.kaxt.startActivityFromClass
-
-class ServiceViewHolder(private val binding: ItemServiceBinding) : androidx.recyclerview.widget.RecyclerView.ViewHolder(binding.root) {
-
-    fun applyService(service: BluetoothGattService) {
-        itemView.setOnClickListener { v ->
-            App.service = service
-            v.context.startActivityFromClass(CharacteristicActivity::class.java)
-        }
-        binding.uuid.text = service.uuid.toString()
-        binding.type.text = DevicePropertiesDescriber.describeServiceType(service)
-        binding.name.text = DevicePropertiesDescriber.getServiceName(itemView.context, service, "unknown")
-    }
-}
